requirements.txt
tgcrypto
python-telegram-bot
sqlalchemy
asyncpg
python-dotenv
loguru
pandas
aiofiles